window1.cs
This code should be placed just above the "public User Control" method that contains the "initializeComponent()" method in it. The code committed is the code that allows this window to be accessed from the outside. The name "ucWindow1" is the name of the "User Control" or "The name of the form (if form is used instead of user control)".